Look for details in API documentation's changelog

+++ OpenAPIClient/Classes/OpenAPIs/Models/SubnetRequestCompound.swift
@@ -16,8 +16,8 @@ public struct SubnetRequestCompound: Codable, JSONEncodable, Hashable {
     static let pkiSubnetIDRule = NumericRule<Int>(minimum: 0, exclusiveMinimum: false, maximum: 65535, exclusiveMaximum: false, multipleOf: nil)
     static let fkiUserIDRule = NumericRule<Int>(minimum: 0, exclusiveMinimum: false, maximum: nil, exclusiveMaximum: false, multipleOf: nil)
     static let fkiApikeyIDRule = NumericRule<Int>(minimum: 0, exclusiveMinimum: false, maximum: nil, exclusiveMaximum: false, multipleOf: nil)
-    static let iSubnetNetworkRule = NumericRule<Int>(minimum: 0, exclusiveMinimum: false, maximum: 4294967295, exclusiveMaximum: false, multipleOf: nil)
-    static let iSubnetMaskRule = NumericRule<Int>(minimum: 0, exclusiveMinimum: false, maximum: 4294967295, exclusiveMaximum: false, multipleOf: nil)
+    static let iSubnetNetworkRule = NumericRule<Int64>(minimum: 0, exclusiveMinimum: false, maximum: 4294967295, exclusiveMaximum: false, multipleOf: nil)
+    static let iSubnetMaskRule = NumericRule<Int64>(minimum: 0, exclusiveMinimum: false, maximum: 4294967295, exclusiveMaximum: false, multipleOf: nil)
     /** The unique ID of the Subnet */
     public var pkiSubnetID: Int?
     /** The unique ID of the User */
@@ -26,11 +26,11 @@ public struct SubnetRequestCompound: Codable, JSONEncodable, Hashable {
     public var fkiApikeyID: Int?
     public var objSubnetDescription: MultilingualSubnetDescription
     /** The network of the Subnet in integer form. For example 8.8.8.0 would be 134744064 */
-    public var iSubnetNetwork: Int
+    public var iSubnetNetwork: Int64
     /** The mask of the Subnet  in integer form. For example 255.255.255.0 would be 4294967040 */
-    public var iSubnetMask: Int
+    public var iSubnetMask: Int64
 
-    public init(pkiSubnetID: Int? = nil, fkiUserID: Int? = nil, fkiApikeyID: Int? = nil, objSubnetDescription: MultilingualSubnetDescription, iSubnetNetwork: Int, iSubnetMask: Int) {
+    public init(pkiSubnetID: Int? = nil, fkiUserID: Int? = nil, fkiApikeyID: Int? = nil, objSubnetDescription: MultilingualSubnetDescription, iSubnetNetwork: Int64, iSubnetMask: Int64) {
         self.pkiSubnetID = pkiSubnetID
         self.fkiUserID = fkiUserID
         self.fkiApikeyID = fkiApikeyID
